refactor(rc-6): split runtime-contracts into oauth-constants + error-sentinels
The file name lib/runtime-contracts.ts was flagged as misleading by the audit (docs/audits/16-code-health.md, RC-6): it held 4 OAuth callback loopback constants (host/port/path/bind-url) plus 2 unrelated string error sentinels with their create/match helpers, not runtime contracts. Split into two purpose-specific modules so each filename matches its content:

- lib/oauth-constants.ts owns OAUTH_CALLBACK_LOOPBACK_HOST/PORT/PATH/BIND_URL.

- lib/error-sentinels.ts owns DEACTIVATED_WORKSPACE_ERROR_CODE and USAGE_REQUEST_TIMEOUT_MESSAGE plus their createDeactivatedWorkspaceError / isDeactivatedWorkspaceErrorMessage / createUsageRequestTimeoutError / isUsageRequestTimeoutMessage helpers.

Updated 4 source importers (index.ts, lib/auth/auth.ts, lib/auth/server.ts, lib/request/fetch-helpers.ts) and 1 test importer (test/doc-parity.test.ts). Split test/runtime-contracts.test.ts along the same axis into test/oauth-constants.test.ts and test/error-sentinels.test.ts so each test file targets one module.

lib/utils.ts was reviewed per RC-6 and intentionally left untouched: 52 lines, 4 generic helpers (isRecord, nowMs, toStringValue, sleep), single responsibility, already accurately named.

No behavior change. typecheck + lint + 2088 tests + build all pass.
