+++ lib/error-sentinels.ts
@@ -1,13 +1,16 @@
 /**
- * Shared runtime constants and sentinel helpers only. This module is pure: it
- * does not perform I/O, persistence, or logging, so centralizing these values
- * does not introduce new Windows lock or token-redaction surfaces.
+ * Sentinel error codes and factory/matcher helpers.
+ *
+ * These sentinels are the runtime contract between throwing code paths (e.g.
+ * request pipeline, usage tracking) and the orchestrator in `index.ts` that
+ * inspects error messages to decide on account rotation, cooldown, and retry
+ * behaviour. Keeping the sentinel string and its create/match helpers in one
+ * module guarantees producers and consumers cannot drift apart.
+ *
+ * This module is pure: it performs no I/O, persistence, or logging, so
+ * centralizing these values does not introduce new Windows lock or
+ * token-redaction surfaces.
  */
-export const OAUTH_CALLBACK_LOOPBACK_HOST = "127.0.0.1";
-export const OAUTH_CALLBACK_PORT = 1455;
-export const OAUTH_CALLBACK_PATH = "/auth/callback";
-export const OAUTH_CALLBACK_BIND_URL = `http://${OAUTH_CALLBACK_LOOPBACK_HOST}:${OAUTH_CALLBACK_PORT}`;
-
 export const DEACTIVATED_WORKSPACE_ERROR_CODE = "deactivated_workspace";
 export const USAGE_REQUEST_TIMEOUT_MESSAGE = "Usage request timed out";
